Extract stats player window controllers
