+++ js/stat-evaluation-player/src/main.ts
@@ -11,18 +11,17 @@ import {
 import type {
   BoostPickupAnimationPickup,
   CanvasRecorderPlugin,
-  CanvasRecorderStatus,
-  CameraSettings,
-  ReplayCameraViewMode,
-  ReplayFreeCameraPreset,
   ReplayTimelineEvent,
   ReplayPlayerState,
-  ReplayPlayerTrack,
   TimelineOverlayPlugin,
 } from "@rlrml/player";
 import { getAppTemplate } from "./appTemplate.ts";
 import { createReplayLoadModal } from "./replayLoadModal.ts";
 import type { ReplayLoadModalController } from "./replayLoadModal.ts";
+import {
+  createCameraControlsController,
+  type CameraControlsController,
+} from "./cameraControls.ts";
 import { createStatModules, getTeamClass, RELATIVE_POSITIONING_MODULE_ID } from "./statModules.ts";
 import type { StatModule, StatModuleContext } from "./statModules.ts";
 import { createBoostPickupFilterController } from "./boostPickupFilters.ts";
@@ -32,15 +31,14 @@ import {
   getConfigAdapterSnapshot,
   type StatsPlayerConfigAdapter,
 } from "./configAdapters.ts";
-import type {
-  PlayerStatsSnapshot,
-  StatsFrame,
-  StatsFrameLookup,
-  TeamStatsSnapshot,
-  StatsTimeline,
-} from "./statsTimeline.ts";
-import { createStatRegistry, type StatDefinition, type StatScopeKind } from "./statRegistry.ts";
-import { getStatDefinitionSearchMatches } from "./statSearch.ts";
+import type { StatsFrameLookup, StatsTimeline } from "./statsTimeline.ts";
+import { createStatRegistry, type StatDefinition } from "./statRegistry.ts";
+import {
+  createStatsWindowsController,
+  formatTime,
+  type RenderStatsWindowsOptions,
+  type StatsWindowsController,
+} from "./statsWindows.ts";
 import {
   filterReplayTimelineEvents,
   formatMechanicKind,
@@ -48,19 +46,25 @@ import {
   mechanicKindToModuleId,
 } from "./timelineMarkers.ts";
 import {
-  buildEventPlaylistItems as buildEventPlaylistSourceItems,
-  getEventPlaylistSelectedSourceIds,
   getEventPlaylistSources as getEventPlaylistSourcesFromTimelineSources,
   getEventTimelineSources as getConfiguredEventTimelineSources,
-  type EventPlaylistSource,
   type EventTimelineSource,
 } from "./eventTimelineSources.ts";
+import {
+  createEventPlaylistWindowController,
+  type EventPlaylistWindowController,
+  type SyncEventPlaylistTimelineOptions,
+} from "./eventPlaylistWindow.ts";
 import {
   formatReplayLoadProgress,
   loadReplayBundleInWorker,
   type ReplayLoadBundle,
   type ReplayLoadProgress,
 } from "./replayLoader.ts";
+import {
+  createRecordingWindowController,
+  type RecordingWindowController,
+} from "./recordingWindow.ts";
 import {
   formatMechanicsReviewBound,
   formatMechanicsReviewClipDetails,
@@ -102,15 +106,12 @@ import {
   type StatsPlayerConfigParamSnapshot,
   type StatsWindowConfig,
   type StatsWindowKind,
-  type TeamScope,
   type WindowPlacementConfig,
 } from "./playerConfig.ts";
-import { playerIdToString } from "./touchOverlay.ts";
 
 const DEFAULT_CAMERA_DISTANCE_SCALE = 2.25;
 const GOAL_WATCH_LEAD_SECONDS = 4;
 const PLAYING_SNAPSHOT_UI_INTERVAL_MS = 100;
-const CAMERA_VIEW_MODES: ReplayCameraViewMode[] = ["free", "follow"];
 
 let replayPlayer: ReplayPlayer | null = null;
 let timelineOverlay: TimelineOverlayPlugin | null = null;
@@ -233,30 +234,6 @@ let touchControlsWindowBody!: HTMLDivElement;
 let statsWindowLayer!: HTMLDivElement;
 let togglePlayback!: HTMLButtonElement;
 let playbackRate!: HTMLSelectElement;
-let attachedPlayer!: HTMLSelectElement;
-let cameraViewFreeButton!: HTMLButtonElement;
-let cameraViewFollowButton!: HTMLButtonElement;
-let cameraViewOverheadButton!: HTMLButtonElement;
-let cameraViewSideButton!: HTMLButtonElement;
-let cameraDistance!: HTMLInputElement;
-let cameraDistanceReadout!: HTMLElement;
-let customCameraSettings!: HTMLInputElement;
-let cameraSettingsControls!: HTMLDivElement;
-let customCameraFov!: HTMLInputElement;
-let customCameraHeight!: HTMLInputElement;
-let customCameraPitch!: HTMLInputElement;
-let customCameraDistance!: HTMLInputElement;
-let customCameraStiffness!: HTMLInputElement;
-let customCameraSwivelSpeed!: HTMLInputElement;
-let customCameraTransitionSpeed!: HTMLInputElement;
-let customCameraFovReadout!: HTMLElement;
-let customCameraHeightReadout!: HTMLElement;
-let customCameraPitchReadout!: HTMLElement;
-let customCameraDistanceReadout!: HTMLElement;
-let customCameraStiffnessReadout!: HTMLElement;
-let customCameraSwivelSpeedReadout!: HTMLElement;
-let customCameraTransitionSpeedReadout!: HTMLElement;
-let ballCam!: HTMLInputElement;
 let moduleSummaryEl!: HTMLDivElement;
 let moduleSettingsEl!: HTMLDivElement;
 let timeReadout!: HTMLElement;
@@ -267,40 +244,22 @@ let statusReadout!: HTMLElement;
 let playersReadout!: HTMLElement;
 let framesReadout!: HTMLElement;
 let eventsReadout!: HTMLElement;
-let cameraProfileReadout!: HTMLElement;
-let cameraFovReadout!: HTMLElement;
-let cameraHeightReadout!: HTMLElement;
-let cameraPitchReadout!: HTMLElement;
-let cameraBaseDistanceReadout!: HTMLElement;
-let cameraStiffnessReadout!: HTMLElement;
 let skipPostGoalTransitions!: HTMLInputElement;
 let replayLoadModal: ReplayLoadModalController | null = null;
 let skipKickoffs!: HTMLInputElement;
 let hitboxWireframes!: HTMLInputElement;
-let recordingFps!: HTMLInputElement;
-let recordingPlaybackRate!: HTMLSelectElement;
-let recordingStart!: HTMLButtonElement;
-let recordingFullReplay!: HTMLButtonElement;
-let recordingStop!: HTMLButtonElement;
-let recordingDownload!: HTMLButtonElement;
-let recordingClear!: HTMLButtonElement;
-let recordingStatus!: HTMLElement;
-let recordingElapsed!: HTMLElement;
-let recordingSize!: HTMLElement;
-let recordingType!: HTMLElement;
 let currentMountCleanup: (() => void) | null = null;
 let statRegistry: StatDefinition[] = createStatRegistry(null);
+let cameraControlsController: CameraControlsController | null = null;
+let recordingWindowController: RecordingWindowController | null = null;
+let statsWindowsController: StatsWindowsController | null = null;
+let eventPlaylistController: EventPlaylistWindowController | null = null;
 let nextWindowZIndex = 30;
-let nextStatsWindowId = 1;
 let boostPadOverlayEnabled = true;
 let loadedReplayName: string | null = null;
-let lastFreeCameraPreset: ReplayFreeCameraPreset | null = null;
 let initialUrlConfig: StatsPlayerConfig | null = null;
 let isApplyingConfig = false;
 let configUrlUpdateTimer: number | null = null;
-let eventPlaylistActiveSourceIds: Set<string> | null = null;
-let eventPlaylistAutoFollow = true;
-let eventPlaylistLastActiveKey: string | null = null;
 
 interface ReplayInputSource {
   name: string;
@@ -322,25 +281,6 @@ const SINGLETON_WINDOW_IDS: SingletonWindowId[] = [
   "touch-controls",
 ];
 
-interface SelectedStatEntry {
-  key: string;
-  statId: string;
-  targetId?: string;
-}
-
-interface StatsWindowState {
-  readonly id: string;
-  readonly kind: StatsWindowKind;
-  readonly entries: SelectedStatEntry[];
-  playerId: string | null;
-  team: TeamScope | null;
-  pickerOpen: boolean;
-  query: string;
-  element: HTMLElement;
-  body: HTMLElement;
-}
-
-const statsWindows = new Map<string, StatsWindowState>();
 let activeMechanicsReview: ActiveMechanicsReview | null = null;
 let mechanicsReviewBoundaryGuard = false;
 
@@ -660,18 +600,27 @@ function applyModuleConfigSnapshot(configs: Record<string, unknown>): void {
   applyConfigAdapterSnapshot(getConfigAdapters(), configs);
 }
 
-function getStatsWindowConfig(statsWindow: StatsWindowState): StatsWindowConfig {
-  return {
-    id: statsWindow.id,
-    kind: statsWindow.kind,
-    placement: readWindowPlacement(statsWindow.element),
-    playerId: statsWindow.playerId,
-    team: statsWindow.team,
-    entries: statsWindow.entries.map((entry) => ({
-      statId: entry.statId,
-      targetId: entry.targetId,
-    })),
-  };
+function getStatsWindowConfigs(): StatsWindowConfig[] {
+  return statsWindowsController?.getConfigs() ?? [];
+}
+
+function renderStatsWindows(
+  frameIndex = replayPlayer?.getState().frameIndex ?? 0,
+  options: RenderStatsWindowsOptions = {},
+): void {
+  statsWindowsController?.render(frameIndex, options);
+}
+
+function createStatsWindow(kind: StatsWindowKind): void {
+  statsWindowsController?.create(kind);
+}
+
+function replaceStatsWindowsFromConfig(configs: readonly StatsWindowConfig[]): void {
+  statsWindowsController?.replaceFromConfig(configs);
+}
+
+function clearStatsWindows(): void {
+  statsWindowsController?.clear();
 }
 
 function getPlaybackConfigSnapshot(): PlayerPlaybackConfig {
@@ -691,19 +640,16 @@ function getCameraConfigSnapshot(): PlayerCameraConfig {
   const state = replayPlayer?.getState();
   return {
     mode: state?.cameraViewMode,
-    freePreset: lastFreeCameraPreset,
+    freePreset: cameraControlsController?.freeCameraPreset ?? null,
     attachedPlayerId: state?.attachedPlayerId,
     distanceScale: state?.cameraDistanceScale,
-    ballCam: state?.ballCamEnabled,
+    ballCam: state?.ballCamEnabled ?? cameraControlsController?.ballCamChecked,
     customSettings: state?.customCameraSettings,
   };
 }
 
 function getRecordingConfigSnapshot(): RecordingConfig {
-  return {
-    fps: Number(recordingFps?.value),
-    playbackRate: Number(recordingPlaybackRate?.value),
-  };
+  return recordingWindowController?.getConfigSnapshot() ?? {};
 }
 
 function getStatsPlayerConfigSnapshot(): StatsPlayerConfig {
@@ -729,7 +675,7 @@ function getStatsPlayerConfigSnapshot(): StatsPlayerConfig {
     },
     recording: getRecordingConfigSnapshot(),
     singletonWindows: getSingletonWindowConfigs(),
-    statsWindows: [...statsWindows.values()].map(getStatsWindowConfig),
+    statsWindows: getStatsWindowConfigs(),
     moduleConfigs: getModuleConfigSnapshot(),
   };
 }
@@ -814,12 +760,7 @@ function applyConfigToStaticControls(config: StatsPlayerConfig): void {
   if (config.playback.rate !== undefined) {
     playbackRate.value = `${config.playback.rate}`;
   }
-  if (config.recording.fps !== undefined) {
-    recordingFps.value = `${config.recording.fps}`;
-  }
-  if (config.recording.playbackRate !== undefined) {
-    recordingPlaybackRate.value = `${config.recording.playbackRate}`;
-  }
+  recordingWindowController?.applyConfig(config.recording);
   applyModuleConfigSnapshot(config.moduleConfigs);
   applyConfigToExistingWindows(config);
   replaceStatsWindowsFromConfig(config.statsWindows);
@@ -863,7 +804,9 @@ function watchGoalReplay(time: number, scorerId: string | null): void {
   if (canFollowScorer) {
     replayPlayer.setAttachedPlayer(scorerId);
     replayPlayer.setCameraViewMode("follow");
-    lastFreeCameraPreset = null;
+    if (cameraControlsController) {
+      cameraControlsController.freeCameraPreset = null;
+    }
   }
 
   skipPostGoalTransitions.checked = false;
@@ -871,6 +814,26 @@ function watchGoalReplay(time: number, scorerId: string | null): void {
   replayPlayer.setState({
     currentTime: Math.max(0, time - GOAL_WATCH_LEAD_SECONDS),
     playing: true,
+    skipPostGoalTransitionsEnabled: false,
+    skipKickoffsEnabled: false,
+  });
+  scheduleConfigUrlUpdate();
+}
+
+function cueGoalReplay(time: number): void {
+  if (!replayPlayer || !Number.isFinite(time)) {
+    return;
+  }
+
+  if (activeMechanicsReview) {
+    activeMechanicsReview.currentClip = null;
+  }
+
+  skipPostGoalTransitions.checked = false;
+  skipKickoffs.checked = false;
+  replayPlayer.setState({
+    currentTime: Math.max(0, time - GOAL_WATCH_LEAD_SECONDS),
+    playing: false,
     skipPostGoalTransitionsEnabled: false,
     skipKickoffsEnabled: false,
   });
@@ -910,7 +873,9 @@ function applyConfigToReplayPlayer(config: StatsPlayerConfig): void {
   replayPlayer.setState(
     getReplayPlayerStatePatchFromConfig(config.playback, config.camera, config),
   );
-  lastFreeCameraPreset = config.camera.freePreset ?? null;
+  if (cameraControlsController) {
+    cameraControlsController.freeCameraPreset = config.camera.freePreset ?? null;
+  }
   if (config.camera.mode === "free" && config.camera.freePreset) {
     replayPlayer.setFreeCameraPreset(config.camera.freePreset);
   }
@@ -1231,246 +1196,24 @@ function getEventSourceColumnCount(sourceCount: number): number {
   return 1;
 }
 
-function getEventPlaylistSources(): EventPlaylistSource[] {
+function getEventPlaylistSourcesForWindow() {
   const ctx = getModuleContext();
   return getEventPlaylistSourcesFromTimelineSources(ctx, getEventTimelineSources(ctx));
 }
 
-function buildEventPlaylistItems(sources: EventPlaylistSource[]) {
-  return buildEventPlaylistSourceItems({
-    sources,
-    activeSourceIds: eventPlaylistActiveSourceIds,
-    replayPlayers: replayPlayer?.replay.players ?? [],
-  });
-}
-
-function setEventPlaylistSourceSelection(
-  sources: EventPlaylistSource[],
-  updater: (selected: Set<string>) => void,
-): void {
-  const selected = getEventPlaylistSelectedSourceIds(sources, eventPlaylistActiveSourceIds);
-  updater(selected);
-  eventPlaylistActiveSourceIds = selected;
-  eventPlaylistLastActiveKey = null;
-  renderEventPlaylistWindow();
-  const state = replayPlayer?.getState();
-  if (state) {
-    syncEventPlaylistTimeline(state);
-  }
-}
-
 function renderEventPlaylistWindow(): void {
-  if (!eventPlaylistWindowBody) {
-    return;
-  }
-
-  eventPlaylistWindowBody.replaceChildren();
-  const sources = getEventPlaylistSources();
-  if (sources.length === 0) {
-    const empty = document.createElement("p");
-    empty.className = "stat-window-empty";
-    empty.textContent = replayPlayer ? "No events loaded." : "Load a replay to see events.";
-    eventPlaylistWindowBody.append(empty);
-    return;
-  }
-
-  const selectedSourceIds = getEventPlaylistSelectedSourceIds(sources, eventPlaylistActiveSourceIds);
-  const items = buildEventPlaylistItems(sources);
-
-  const toolbar = document.createElement("div");
-  toolbar.className = "event-playlist-toolbar";
-
-  const filters = document.createElement("details");
-  filters.className = "event-playlist-filter";
-  filters.dataset.noDrag = "true";
-
-  const summary = document.createElement("summary");
-  summary.textContent = `Filters ${selectedSourceIds.size}/${sources.length}`;
-  filters.append(summary);
-
-  const filterPanel = document.createElement("div");
-  filterPanel.className = "event-playlist-filter-panel";
-
-  const actions = document.createElement("div");
-  actions.className = "event-playlist-filter-actions";
-
-  const allButton = document.createElement("button");
-  allButton.type = "button";
-  allButton.textContent = "All";
-  allButton.addEventListener("click", () => {
-    eventPlaylistActiveSourceIds = new Set(sources.map((source) => source.id));
-    eventPlaylistLastActiveKey = null;
-    renderEventPlaylistWindow();
-    const state = replayPlayer?.getState();
-    if (state) syncEventPlaylistTimeline(state);
-  });
-
-  const noneButton = document.createElement("button");
-  noneButton.type = "button";
-  noneButton.textContent = "None";
-  noneButton.addEventListener("click", () => {
-    eventPlaylistActiveSourceIds = new Set();
-    eventPlaylistLastActiveKey = null;
-    renderEventPlaylistWindow();
-  });
-
-  actions.append(allButton, noneButton);
-  filterPanel.append(actions);
-
-  const sourcesByGroup = new Map<string, EventPlaylistSource[]>();
-  for (const source of sources) {
-    const group = sourcesByGroup.get(source.group) ?? [];
-    group.push(source);
-    sourcesByGroup.set(source.group, group);
-  }
-
-  for (const [group, groupSources] of sourcesByGroup) {
-    const groupEl = document.createElement("section");
-    groupEl.className = "event-playlist-filter-group";
-    const heading = document.createElement("h3");
-    heading.textContent = group;
-    groupEl.append(heading);
-
-    for (const source of groupSources) {
-      const label = document.createElement("label");
-      label.className = "toggle event-playlist-filter-option";
-
-      const input = document.createElement("input");
-      input.type = "checkbox";
-      input.checked = selectedSourceIds.has(source.id);
-      input.addEventListener("change", () => {
-        setEventPlaylistSourceSelection(sources, (selected) => {
-          if (input.checked) {
-            selected.add(source.id);
-          } else {
-            selected.delete(source.id);
-          }
-        });
-      });
-
-      const text = document.createElement("span");
-      text.textContent = `${source.label} (${source.events.length})`;
-      label.append(input, text);
-      groupEl.append(label);
-    }
-
-    filterPanel.append(groupEl);
-  }
-
-  filters.append(filterPanel);
-
-  const followLabel = document.createElement("label");
-  followLabel.className = "toggle event-playlist-follow";
-  const followInput = document.createElement("input");
-  followInput.type = "checkbox";
-  followInput.checked = eventPlaylistAutoFollow;
-  followInput.addEventListener("change", () => {
-    eventPlaylistAutoFollow = followInput.checked;
-    const state = replayPlayer?.getState();
-    if (state) syncEventPlaylistTimeline(state, { forceScroll: true });
-  });
-  const followText = document.createElement("span");
-  followText.textContent = "Auto-follow";
-  followLabel.append(followInput, followText);
-
-  toolbar.append(filters, followLabel);
-
-  const list = document.createElement("div");
-  list.className = "event-playlist-list";
-  list.dataset.noDrag = "true";
-
-  if (items.length === 0) {
-    const empty = document.createElement("p");
-    empty.className = "stat-window-empty";
-    empty.textContent = "No event types selected.";
-    list.append(empty);
-  } else {
-    for (const item of items) {
-      const button = document.createElement("button");
-      button.type = "button";
-      button.className = "event-playlist-item";
-      button.dataset.eventKey = item.key;
-      button.dataset.eventTime = `${item.event.time}`;
-      button.style.setProperty("--event-color", item.color);
-      button.addEventListener("click", () => {
-        cueTimelineEvent(item.event);
-      });
-
-      const time = document.createElement("span");
-      time.className = "event-playlist-time";
-      time.textContent = formatTime(item.event.time);
-
-      const main = document.createElement("span");
-      main.className = "event-playlist-main";
-      const label = document.createElement("strong");
-      label.textContent = item.event.label ?? item.sourceLabel;
-      const meta = document.createElement("span");
-      meta.textContent = [
-        item.event.playerName ?? null,
-        item.event.frame !== undefined ? `frame ${item.event.frame}` : null,
-        item.sourceLabel,
-      ]
-        .filter((part): part is string => Boolean(part))
-        .join(" · ");
-      main.append(label, meta);
-
-      button.append(time, main);
-      list.append(button);
-    }
-  }
-
-  eventPlaylistWindowBody.append(toolbar, list);
-}
-
-function getEventPlaylistActiveItem(list: HTMLElement, currentTime: number): HTMLElement | null {
-  const items = [...list.querySelectorAll<HTMLElement>(".event-playlist-item")];
-  if (items.length === 0) {
-    return null;
-  }
-
-  let bestItem = items[0] ?? null;
-  let bestDistance = Number.POSITIVE_INFINITY;
-  for (const item of items) {
-    const time = Number(item.dataset.eventTime);
-    if (!Number.isFinite(time)) {
-      continue;
-    }
-    const distance = Math.abs(time - currentTime);
-    if (distance < bestDistance) {
-      bestDistance = distance;
-      bestItem = item;
-    }
-  }
-  return bestItem;
+  eventPlaylistController?.render();
 }
 
 function syncEventPlaylistTimeline(
   state: ReplayPlayerState,
-  options: { forceScroll?: boolean } = {},
+  options: SyncEventPlaylistTimelineOptions = {},
 ): void {
-  const list = eventPlaylistWindowBody?.querySelector<HTMLElement>(".event-playlist-list");
-  if (!list) {
-    return;
-  }
+  eventPlaylistController?.syncTimeline(state, options);
+}
 
-  const activeItem = getEventPlaylistActiveItem(list, state.currentTime);
-  const activeKey = activeItem?.dataset.eventKey ?? null;
-  if (activeKey === eventPlaylistLastActiveKey && !options.forceScroll) {
-    return;
-  }
-
-  list.querySelectorAll<HTMLElement>(".event-playlist-item[data-active='true']").forEach((item) => {
-    item.dataset.active = "false";
-  });
-
-  if (activeItem) {
-    activeItem.dataset.active = "true";
-    if (eventPlaylistAutoFollow || options.forceScroll) {
-      activeItem.scrollIntoView({ block: "nearest" });
-    }
-  }
-
-  eventPlaylistLastActiveKey = activeKey;
+function resetEventPlaylistWindow(): void {
+  eventPlaylistController?.reset();
 }
 
 function createMechanicsReviewReplaySource(
@@ -1982,7 +1725,9 @@ async function activateMechanicsReviewItem(index: number): Promise<void> {
     if (playerId && replayPlayer?.replay.players.some((player) => player.id === playerId)) {
       replayPlayer.setAttachedPlayer(playerId);
       replayPlayer.setCameraViewMode("follow");
-      lastFreeCameraPreset = null;
+      if (cameraControlsController) {
+        cameraControlsController.freeCameraPreset = null;
+      }
     }
 
     skipPostGoalTransitions.checked = false;
@@ -2218,7 +1963,7 @@ function renderScoreboard(frameIndex = replayPlayer?.getState().frameIndex ?? 0)
   }
 
   scoreboardWindowBody.replaceChildren();
-  const frame = getCurrentStatsFrame(frameIndex);
+  const frame = statsFrameLookup ? getStatsFrameForReplayFrame(statsFrameLookup, frameIndex) : null;
   const replay = replayPlayer?.replay ?? null;
   if (!frame || !replay) {
     const empty = document.createElement("p");
@@ -2270,1114 +2015,17 @@ function renderTouchControlsWindow(): void {
   }
 }
 
-function getStatById(statId: string): StatDefinition | null {
-  return statRegistry.find((definition) => definition.id === statId) ?? null;
-}
-
-function getCurrentStatsFrame(frameIndex: number): StatsFrame | null {
-  return statsFrameLookup ? getStatsFrameForReplayFrame(statsFrameLookup, frameIndex) : null;
-}
-
-function getTeamSnapshot(frame: StatsFrame, team: TeamScope): TeamStatsSnapshot | null {
-  return team === "blue" ? (frame.team_zero ?? null) : (frame.team_one ?? null);
-}
-
-function getTeamLabel(team: TeamScope): string {
-  return team === "blue" ? "Blue" : "Orange";
-}
-
-function getPlayerTeamClass(playerId: string | null | undefined): string | null {
-  const player = replayPlayer?.replay.players.find((candidate) => candidate.id === playerId);
-  return player ? getTeamClass(player.isTeamZero) : null;
-}
-
-function getTeamScopeClass(team: TeamScope): string {
-  return getTeamClass(team === "blue");
-}
-
-function appendGroupedPlayerOptions(
-  select: HTMLSelectElement,
-  selectedPlayerId: string | null | undefined,
-): void {
-  const players = replayPlayer?.replay.players ?? [];
-  for (const team of ["blue", "orange"] as const) {
-    const teamPlayers = players.filter((player) => player.isTeamZero === (team === "blue"));
-    if (teamPlayers.length === 0) {
-      continue;
-    }
-
-    const group = document.createElement("optgroup");
-    group.label = `${getTeamLabel(team)} team`;
-    for (const player of teamPlayers) {
-      group.append(
-        new Option(
-          player.name,
-          player.id,
-          player.id === selectedPlayerId,
-          player.id === selectedPlayerId,
-        ),
-      );
-    }
-    select.append(group);
-  }
-}
-
-function getStatsWindowScopeTeamClass(statsWindow: StatsWindowState): string | null {
-  if (statsWindow.kind === "player") {
-    return getPlayerTeamClass(statsWindow.playerId);
-  }
-  if (statsWindow.kind === "team") {
-    return getTeamScopeClass(statsWindow.team ?? "blue");
-  }
-  return null;
-}
-
-function getStatTargetTeamClass(
-  definition: StatDefinition,
-  targetId: string | undefined,
-): string | null {
-  if (definition.scope === "player") {
-    return getPlayerTeamClass(targetId);
-  }
-  return getTeamScopeClass(targetId === "orange" ? "orange" : "blue");
-}
-
-function getStatsWindowTitle(kind: StatsWindowKind): string {
-  switch (kind) {
-    case "player":
-      return "Player stats";
-    case "team":
-      return "Team stats";
-    case "all-players":
-      return "All players stats";
-    case "all-teams":
-      return "All teams stats";
-    case "goals-overview":
-      return "Goal labels";
-    case "ad-hoc":
-      return "Ad hoc stats";
-  }
-}
-
-function hasStatsWindowScopeSelector(kind: StatsWindowKind): boolean {
-  return kind === "player" || kind === "team";
-}
-
-function hasStatsWindowStatPicker(kind: StatsWindowKind): boolean {
-  return kind !== "goals-overview";
-}
-
-function getStatsWindowAllowedScope(kind: StatsWindowKind): StatScopeKind | null {
-  switch (kind) {
-    case "player":
-    case "all-players":
-      return "player";
-    case "team":
-    case "all-teams":
-      return "team";
-    case "goals-overview":
-      return null;
-    case "ad-hoc":
-      return null;
-  }
-}
-
-function getStatsWindowDefaultPosition(): { x: number; y: number } {
-  const offset = statsWindows.size * 18;
-  return {
-    x: Math.max(12, Math.min(window.innerWidth - 360, 96 + offset)),
-    y: Math.max(64, Math.min(window.innerHeight - 240, 96 + offset)),
-  };
-}
-
-function renderStatsWindows(
-  frameIndex = replayPlayer?.getState().frameIndex ?? 0,
-  options: { preserveOpenPickers?: boolean } = {},
-): void {
-  for (const statsWindow of statsWindows.values()) {
-    if (
-      options.preserveOpenPickers &&
-      (statsWindow.pickerOpen || statsWindow.element.contains(document.activeElement))
-    ) {
-      continue;
-    }
-    renderStatsWindow(statsWindow, frameIndex);
-  }
-}
-
-function createStatsWindow(kind: StatsWindowKind, config?: StatsWindowConfig): StatsWindowState {
-  const id = config?.id ?? `stats-${nextStatsWindowId++}`;
-  const idNumber = Number.parseInt(id.replace(/^stats-/, ""), 10);
-  if (Number.isFinite(idNumber)) {
-    nextStatsWindowId = Math.max(nextStatsWindowId, idNumber + 1);
-  }
-  const { x, y } = getStatsWindowDefaultPosition();
-  const element = document.createElement("section");
-  element.className = "stats-window";
-  element.dataset.windowId = id;
-  element.style.setProperty("--window-x", `${x}px`);
-  element.style.setProperty("--window-y", `${y}px`);
-  if (config) {
-    applyWindowPlacement(element, config.placement);
-  }
-
-  const header = document.createElement("header");
-  header.className = "stats-window-header";
-
-  const actions = document.createElement("div");
-  actions.className = "stats-window-actions";
-  const hideButton = document.createElement("button");
-  hideButton.type = "button";
-  hideButton.className = "stats-window-action";
-  hideButton.textContent = "Hide";
-  actions.append(hideButton);
-  if (hasStatsWindowScopeSelector(kind)) {
-    header.classList.add("stats-window-header-actions-only");
-    header.append(actions);
-  } else {
-    const title = document.createElement("h2");
-    title.textContent = getStatsWindowTitle(kind);
-    header.append(title, actions);
-  }
-
-  const body = document.createElement("div");
-  body.className = "stats-window-body";
-  element.append(header, body);
-  statsWindowLayer.append(element);
-
-  const state: StatsWindowState = {
-    id,
-    kind,
-    entries:
-      config?.entries.map((entry) => ({
-        key: `${id}:${entry.statId}:${entry.targetId ?? "scope"}`,
-        statId: entry.statId,
-        targetId: entry.targetId,
-      })) ?? [],
-    playerId: config?.playerId ?? replayPlayer?.replay.players[0]?.id ?? null,
-    team: config?.team ?? "blue",
-    pickerOpen: false,
-    query: "",
-    element,
-    body,
-  };
-
-  hideButton.addEventListener("click", () => {
-    element.hidden = true;
-    scheduleConfigUrlUpdate();
-  });
-
-  statsWindows.set(id, state);
-  if (!config) {
-    bringWindowToFront(element);
-  }
-  setLauncherOpen(false);
-  renderStatsWindow(state);
-  scheduleConfigUrlUpdate();
-  return state;
-}
-
-function replaceStatsWindowsFromConfig(configs: readonly StatsWindowConfig[]): void {
-  for (const statsWindow of statsWindows.values()) {
-    statsWindow.element.remove();
-  }
-  statsWindows.clear();
-  nextStatsWindowId = 1;
-  for (const config of configs) {
-    createStatsWindow(config.kind, config);
-  }
-}
-
-function renderStatsWindow(
-  statsWindow: StatsWindowState,
-  frameIndex = replayPlayer?.getState().frameIndex ?? 0,
-): void {
-  const activeElement = document.activeElement;
-  const searchFocused =
-    activeElement instanceof HTMLInputElement &&
-    activeElement.dataset.statsWindowSearch === statsWindow.id;
-  const searchSelectionStart = searchFocused ? activeElement.selectionStart : null;
-  const searchSelectionEnd = searchFocused ? activeElement.selectionEnd : null;
-  const searchSelectionDirection = searchFocused ? activeElement.selectionDirection : null;
-
-  statsWindow.body.replaceChildren();
-
-  renderStatsWindowScope(statsWindow);
-  if (hasStatsWindowStatPicker(statsWindow.kind)) {
-    renderStatsWindowAddControl(statsWindow);
-    renderStatsWindowPicker(statsWindow);
-  }
-  renderStatsWindowEntries(statsWindow, frameIndex);
-
-  if (searchFocused) {
-    const searchInput = statsWindow.body.querySelector<HTMLInputElement>(
-      `input[data-stats-window-search="${statsWindow.id}"]`,
-    );
-    searchInput?.focus({ preventScroll: true });
-    if (searchInput && searchSelectionStart !== null && searchSelectionEnd !== null) {
-      searchInput.setSelectionRange(
-        searchSelectionStart,
-        searchSelectionEnd,
-        searchSelectionDirection ?? "none",
-      );
-    }
-  }
-}
-
-function renderStatsWindowScope(statsWindow: StatsWindowState): void {
-  if (statsWindow.kind !== "player" && statsWindow.kind !== "team") {
-    return;
-  }
-
-  const row = document.createElement("div");
-  row.className = "stats-window-scope-row";
-
-  const select = document.createElement("select");
-  select.className = "stats-window-scope-select";
-  const teamClass = getStatsWindowScopeTeamClass(statsWindow);
-  if (teamClass) {
-    select.classList.add(teamClass);
-  }
-  select.setAttribute(
-    "aria-label",
-    statsWindow.kind === "player" ? "Player stats target" : "Team stats target",
-  );
-  if (statsWindow.kind === "player") {
-    appendGroupedPlayerOptions(select, statsWindow.playerId);
-    select.value = statsWindow.playerId ?? "";
-    select.addEventListener("change", () => {
-      statsWindow.playerId = select.value || null;
-      renderStatsWindow(statsWindow);
-      scheduleConfigUrlUpdate();
-    });
-  } else {
-    select.append(
-      new Option("Blue", "blue", statsWindow.team === "blue", statsWindow.team === "blue"),
-      new Option("Orange", "orange", statsWindow.team === "orange", statsWindow.team === "orange"),
-    );
-    select.value = statsWindow.team ?? "blue";
-    select.addEventListener("change", () => {
-      statsWindow.team = select.value === "orange" ? "orange" : "blue";
-      renderStatsWindow(statsWindow);
-      scheduleConfigUrlUpdate();
-    });
-  }
-
-  row.append(select);
-  statsWindow.body.append(row);
-}
-
-function renderStatsWindowAddControl(statsWindow: StatsWindowState): void {
-  const button = document.createElement("button");
-  button.type = "button";
-  button.className = "stats-window-add-button";
-  button.textContent = "+";
-  button.title = "Add stat";
-  button.setAttribute("aria-label", "Add stat");
-  button.setAttribute("aria-expanded", String(statsWindow.pickerOpen));
-  activateButton(button, () => {
-    statsWindow.pickerOpen = !statsWindow.pickerOpen;
-    renderStatsWindow(statsWindow);
-  });
-
-  if (hasStatsWindowScopeSelector(statsWindow.kind)) {
-    const scopeRow = statsWindow.body.querySelector(".stats-window-scope-row");
-    scopeRow?.append(button);
-    return;
-  }
-
-  const toolbar = document.createElement("div");
-  toolbar.className = "stats-window-toolbar";
-  toolbar.append(button);
-  statsWindow.body.append(toolbar);
-}
-
-function activateButton(button: HTMLButtonElement, callback: () => void): void {
-  let pointerActivated = false;
-  button.addEventListener("pointerdown", (event) => {
-    if (button.disabled) {
-      return;
-    }
-    pointerActivated = true;
-    event.preventDefault();
-    callback();
-  });
-  button.addEventListener("click", () => {
-    if (pointerActivated) {
-      pointerActivated = false;
-      return;
-    }
-    if (!button.disabled) {
-      callback();
-    }
-  });
-}
-
-function renderStatsWindowPicker(statsWindow: StatsWindowState): void {
-  const picker = document.createElement("div");
-  picker.className = "stats-window-picker";
-  picker.hidden = !statsWindow.pickerOpen;
-  if (picker.hidden) {
-    statsWindow.body.append(picker);
-    return;
-  }
-
-  const allowedScope = getStatsWindowAllowedScope(statsWindow.kind);
-  const queryInput = document.createElement("input");
-  queryInput.type = "search";
-  queryInput.placeholder = "Search stats";
-  queryInput.value = statsWindow.query;
-  queryInput.dataset.statsWindowSearch = statsWindow.id;
-
-  const list = document.createElement("div");
-  list.className = "stats-window-picker-list";
-  queryInput.addEventListener("input", () => {
-    statsWindow.query = queryInput.value;
-    renderStatsWindowPickerList(statsWindow, list, allowedScope);
-  });
-
-  renderStatsWindowPickerList(statsWindow, list, allowedScope);
-
-  picker.append(queryInput, list);
-  statsWindow.body.append(picker);
-}
-
-function renderStatsWindowPickerList(
-  statsWindow: StatsWindowState,
-  list: HTMLElement,
-  allowedScope: StatScopeKind | null,
-): void {
-  list.replaceChildren();
-
-  const scopeDefinitions = allowedScope
-    ? statRegistry.filter((definition) => definition.scope === allowedScope)
-    : statRegistry;
-  const definitions = getStatDefinitionSearchMatches(scopeDefinitions, statsWindow.query);
-
-  const groupByCategory = new Map<string, StatDefinition[]>();
-  for (const definition of definitions) {
-    const group = groupByCategory.get(definition.category) ?? [];
-    group.push(definition);
-    groupByCategory.set(definition.category, group);
-  }
-
-  for (const [category, group] of groupByCategory) {
-    if (group.length < 2) continue;
-    const addGroup = document.createElement("button");
-    addGroup.type = "button";
-    addGroup.className = "stats-window-picker-item";
-    addGroup.innerHTML = `<span>Add all ${category}</span><strong>${group.length}</strong>`;
-    activateButton(addGroup, () => {
-      for (const definition of group) {
-        addStatToWindow(statsWindow, definition);
-      }
-      renderStatsWindow(statsWindow);
-      scheduleConfigUrlUpdate();
-    });
-    list.append(addGroup);
-  }
-
-  for (const definition of definitions) {
-    const item = document.createElement("button");
-    item.type = "button";
-    item.className = "stats-window-picker-item";
-    item.innerHTML = `<span>${definition.label}</span><strong>${definition.scope}</strong>`;
-    item.disabled =
-      statsWindow.kind !== "ad-hoc" &&
-      statsWindow.entries.some((entry) => entry.statId === definition.id);
-    activateButton(item, () => {
-      addStatToWindow(statsWindow, definition);
-      renderStatsWindow(statsWindow);
-      scheduleConfigUrlUpdate();
-    });
-    list.append(item);
-  }
-
-  if (definitions.length === 0) {
-    const empty = document.createElement("p");
-    empty.className = "stat-window-empty";
-    empty.textContent = statRegistry.length === 0 ? "No stats available." : "No matching stats.";
-    list.append(empty);
-  }
-}
-
-function addStatToWindow(statsWindow: StatsWindowState, definition: StatDefinition): void {
-  const targetId = statsWindow.kind === "ad-hoc" ? getDefaultAdHocTargetId(definition) : undefined;
-  if (
-    statsWindow.entries.some(
-      (entry) => entry.statId === definition.id && entry.targetId === targetId,
-    )
-  ) {
-    return;
-  }
-  statsWindow.entries.push({
-    key: `${statsWindow.id}:${definition.id}:${targetId ?? "scope"}`,
-    statId: definition.id,
-    targetId,
-  });
-}
-
-function getDefaultAdHocTargetId(definition: StatDefinition): string {
-  if (definition.scope === "player") {
-    return replayPlayer?.replay.players[0]?.id ?? "";
-  }
-  return "blue";
-}
-
-function removeStatFromWindow(statsWindow: StatsWindowState, entryKey: string): void {
-  const index = statsWindow.entries.findIndex((entry) => entry.key === entryKey);
-  if (index >= 0) {
-    statsWindow.entries.splice(index, 1);
-  }
-}
-
-function renderStatsWindowEntries(statsWindow: StatsWindowState, frameIndex: number): void {
-  if (statsWindow.kind === "goals-overview") {
-    renderGoalLabelsOverview(statsWindow);
-    return;
-  }
-
-  const allowedScope = getStatsWindowAllowedScope(statsWindow.kind);
-  const entries = statsWindow.entries
-    .map((entry) => ({ entry, definition: getStatById(entry.statId) }))
-    .filter(
-      (item): item is { entry: SelectedStatEntry; definition: StatDefinition } =>
-        item.definition !== null && (!allowedScope || item.definition.scope === allowedScope),
-    );
-
-  if (entries.length === 0) {
-    const empty = document.createElement("p");
-    empty.className = "stat-window-empty";
-    empty.textContent = "No stats added.";
-    statsWindow.body.append(empty);
-    return;
-  }
-
-  const frame = getCurrentStatsFrame(frameIndex);
-  if (!frame) {
-    const empty = document.createElement("p");
-    empty.className = "stat-window-empty";
-    empty.textContent = "Load a replay to show stats.";
-    statsWindow.body.append(empty);
-    return;
-  }
-
-  if (statsWindow.kind === "all-players") {
-    renderAllPlayersStats(statsWindow, frame, entries);
-    return;
-  }
-  if (statsWindow.kind === "all-teams") {
-    renderAllTeamsStats(statsWindow, frame, entries);
-    return;
-  }
-  if (statsWindow.kind === "player") {
-    const player = statsWindow.playerId
-      ? (frame.players.find(
-          (candidate) => playerIdToString(candidate.player_id) === statsWindow.playerId,
-        ) ?? null)
-      : null;
-    renderScopedStatList(statsWindow, player, entries);
-    return;
-  }
-  if (statsWindow.kind === "team") {
-    renderScopedStatList(statsWindow, getTeamSnapshot(frame, statsWindow.team ?? "blue"), entries);
-    return;
-  }
-  if (statsWindow.kind === "ad-hoc") {
-    renderAdHocStats(statsWindow, frame, entries);
-  }
-}
-
-function renderGoalLabelsOverview(statsWindow: StatsWindowState): void {
-  const timeline = statsTimeline;
-  const replay = replayPlayer?.replay ?? null;
-  if (!timeline || !replay) {
-    appendStatsWindowEmpty(statsWindow, "Load a replay to show goal labels.");
-    return;
-  }
-
-  const goalContexts = [...(timeline.events.goal_context ?? [])].sort(
-    (left, right) => left.time - right.time,
-  );
-  const tagsByGoalIndex = new Map<number, typeof timeline.events.goal_tags>();
-  for (const tag of timeline.events.goal_tags ?? []) {
-    const group = tagsByGoalIndex.get(tag.goal_index) ?? [];
-    group.push(tag);
-    tagsByGoalIndex.set(tag.goal_index, group);
-  }
-  for (const group of tagsByGoalIndex.values()) {
-    group.sort(
-      (left, right) => left.kind.localeCompare(right.kind) || right.confidence - left.confidence,
-    );
-  }
-
-  const goalIndexes = new Set<number>(goalContexts.map((_, index) => index));
-  for (const index of tagsByGoalIndex.keys()) {
-    goalIndexes.add(index);
-  }
-  const orderedGoalIndexes = [...goalIndexes].sort((left, right) => left - right);
-  if (orderedGoalIndexes.length === 0) {
-    appendStatsWindowEmpty(statsWindow, "No goals loaded.");
-    return;
-  }
-
-  const list = document.createElement("div");
-  list.className = "goal-label-list";
-  for (const goalIndex of orderedGoalIndexes) {
-    const context = goalContexts[goalIndex] ?? null;
-    const tags = tagsByGoalIndex.get(goalIndex) ?? [];
-    const firstTag = tags[0] ?? null;
-    const time = context?.time ?? firstTag?.time ?? 0;
-    const scorer = context?.scorer ?? firstTag?.scorer ?? null;
-    const scorerId = scorer ? playerIdToString(scorer) : null;
-    const scorerName = scorer
-      ? (replay.players.find((player) => player.id === scorerId)?.name ?? scorerId)
-      : "Unknown scorer";
-    const isTeamZero = context?.scoring_team_is_team_0 ?? firstTag?.scoring_team_is_team_0 ?? null;
-
-    const item = document.createElement("section");
-    item.className = "goal-label-item";
-    if (isTeamZero !== null) {
-      item.classList.add(getTeamClass(isTeamZero));
-    }
-
-    const header = document.createElement("header");
-    const title = document.createElement("h3");
-    title.textContent = `Goal ${goalIndex + 1}`;
-    const meta = document.createElement("span");
-    meta.textContent = `${formatTime(time)} · ${scorerName}`;
-    header.append(title, meta);
-
-    const labels = document.createElement("div");
-    labels.className = "goal-label-tags";
-    if (tags.length === 0) {
-      const empty = document.createElement("span");
-      empty.className = "goal-label-tag goal-label-tag-empty";
-      empty.textContent = "Unlabeled";
-      labels.append(empty);
-    } else {
-      for (const tag of tags) {
-        const chip = document.createElement("span");
-        chip.className = "goal-label-tag";
-        chip.textContent = `${formatMechanicKind(tag.kind)} ${Math.round(tag.confidence * 100)}%`;
-        labels.append(chip);
-      }
-    }
-
-    const actions = document.createElement("div");
-    actions.className = "goal-label-actions";
-    const watch = document.createElement("button");
-    watch.type = "button";
-    watch.className = "goal-label-watch";
-    watch.textContent = "Watch";
-    watch.addEventListener("click", () => {
-      watchGoalReplay(time, scorerId);
-    });
-    const jump = document.createElement("button");
-    jump.type = "button";
-    jump.textContent = "Cue";
-    jump.addEventListener("click", () => {
-      replayPlayer?.setState({
-        currentTime: Math.max(0, time - GOAL_WATCH_LEAD_SECONDS),
-        playing: false,
-        skipPostGoalTransitionsEnabled: false,
-        skipKickoffsEnabled: false,
-      });
-      skipPostGoalTransitions.checked = false;
-      skipKickoffs.checked = false;
-      scheduleConfigUrlUpdate();
-    });
-    actions.append(watch, jump);
-
-    item.append(header, labels, actions);
-    list.append(item);
-  }
-  statsWindow.body.append(list);
-}
-
-function appendStatsWindowEmpty(statsWindow: StatsWindowState, message: string): void {
-  const empty = document.createElement("p");
-  empty.className = "stat-window-empty";
-  empty.textContent = message;
-  statsWindow.body.append(empty);
-}
-
-function formatTime(seconds: number): string {
-  if (!Number.isFinite(seconds)) {
-    return "--";
-  }
-  const minutes = Math.floor(Math.max(0, seconds) / 60);
-  const remainingSeconds = Math.max(0, seconds) - minutes * 60;
-  return `${minutes}:${remainingSeconds.toFixed(1).padStart(4, "0")}`;
-}
-
-function renderScopedStatList(
-  statsWindow: StatsWindowState,
-  target: PlayerStatsSnapshot | TeamStatsSnapshot | null,
-  entries: Array<{ entry: SelectedStatEntry; definition: StatDefinition }>,
-): void {
-  const list = document.createElement("div");
-  list.className = "stats-window-stat-list";
-  for (const { entry, definition } of entries) {
-    list.append(
-      renderStatRow(
-        statsWindow,
-        entry,
-        definition,
-        target ? definition.format(definition.read(target)) : "--",
-      ),
-    );
-  }
-  statsWindow.body.append(list);
-}
-
-function renderAllPlayersStats(
-  statsWindow: StatsWindowState,
-  frame: StatsFrame,
-  entries: Array<{ entry: SelectedStatEntry; definition: StatDefinition }>,
-): void {
-  const list = document.createElement("div");
-  list.className = "stats-window-team-list";
-  for (const team of ["blue", "orange"] as const) {
-    const players = frame.players.filter((player) => player.is_team_0 === (team === "blue"));
-    if (players.length === 0) {
-      continue;
-    }
-
-    const teamSection = document.createElement("section");
-    teamSection.className = `stats-window-team-group ${getTeamScopeClass(team)}`;
-
-    const teamHeader = document.createElement("header");
-    teamHeader.className = "stats-window-team-header";
-    const teamTitle = document.createElement("h3");
-    teamTitle.textContent = `${getTeamLabel(team)} team`;
-    const teamMeta = document.createElement("span");
-    teamMeta.textContent = `${players.length} player${players.length === 1 ? "" : "s"}`;
-    teamHeader.append(teamTitle, teamMeta);
-    teamSection.append(teamHeader);
-
-    const playerList = document.createElement("div");
-    playerList.className = "stats-window-entity-list";
-    for (const player of players) {
-      const section = document.createElement("section");
-      section.className = `stats-window-entity ${getTeamClass(player.is_team_0)}`;
-      const title = document.createElement("h4");
-      title.className = "stats-window-entity-title";
-      title.textContent = player.name;
-      section.append(title);
-      for (const { entry, definition } of entries) {
-        section.append(
-          renderStatRow(statsWindow, entry, definition, definition.format(definition.read(player))),
-        );
-      }
-      playerList.append(section);
-    }
-    teamSection.append(playerList);
-    list.append(teamSection);
-  }
-  statsWindow.body.append(list);
-}
-
-function renderAllTeamsStats(
-  statsWindow: StatsWindowState,
-  frame: StatsFrame,
-  entries: Array<{ entry: SelectedStatEntry; definition: StatDefinition }>,
-): void {
-  const list = document.createElement("div");
-  list.className = "stats-window-entity-list";
-  for (const team of ["blue", "orange"] as const) {
-    const snapshot = getTeamSnapshot(frame, team);
-    const section = document.createElement("section");
-    section.className = `stats-window-entity ${getTeamScopeClass(team)}`;
-    const title = document.createElement("h3");
-    title.className = "stats-window-entity-title";
-    title.textContent = getTeamLabel(team);
-    section.append(title);
-    for (const { entry, definition } of entries) {
-      section.append(
-        renderStatRow(
-          statsWindow,
-          entry,
-          definition,
-          snapshot ? definition.format(definition.read(snapshot)) : "--",
-        ),
-      );
-    }
-    list.append(section);
-  }
-  statsWindow.body.append(list);
-}
-
-function renderAdHocStats(
-  statsWindow: StatsWindowState,
-  frame: StatsFrame,
-  entries: Array<{ entry: SelectedStatEntry; definition: StatDefinition }>,
-): void {
-  const list = document.createElement("div");
-  list.className = "stats-window-stat-list";
-  for (const { entry, definition } of entries) {
-    const target = getAdHocTarget(frame, definition, entry.targetId);
-    list.append(
-      renderStatRow(
-        statsWindow,
-        entry,
-        definition,
-        target ? definition.format(definition.read(target)) : "--",
-      ),
-    );
-  }
-  statsWindow.body.append(list);
-}
-
-function getAdHocTarget(
-  frame: StatsFrame,
-  definition: StatDefinition,
-  targetId: string | undefined,
-): PlayerStatsSnapshot | TeamStatsSnapshot | null {
-  if (definition.scope === "player") {
-    return (
-      frame.players.find((player) => playerIdToString(player.player_id) === targetId) ??
-      frame.players[0] ??
-      null
-    );
-  }
-  return getTeamSnapshot(frame, targetId === "orange" ? "orange" : "blue");
-}
-
-function renderStatRow(
-  statsWindow: StatsWindowState,
-  entry: SelectedStatEntry,
-  definition: StatDefinition,
-  value: string,
-): HTMLElement {
-  const row = document.createElement("div");
-  row.className = "stats-window-stat-row";
-  const name = document.createElement("span");
-  name.className = "stats-window-stat-name";
-  name.textContent = definition.label;
-  if (statsWindow.kind === "ad-hoc") {
-    const targetSelect = document.createElement("select");
-    targetSelect.className = "stats-window-stat-target";
-    const teamClass = getStatTargetTeamClass(definition, entry.targetId);
-    if (teamClass) {
-      targetSelect.classList.add(teamClass);
-    }
-    if (definition.scope === "player") {
-      appendGroupedPlayerOptions(targetSelect, entry.targetId);
-    } else {
-      targetSelect.append(
-        new Option("Blue", "blue", entry.targetId === "blue", entry.targetId === "blue"),
-        new Option("Orange", "orange", entry.targetId === "orange", entry.targetId === "orange"),
-      );
-    }
-    targetSelect.value = entry.targetId ?? "";
-    targetSelect.addEventListener("change", () => {
-      const nextTargetId = targetSelect.value;
-      if (
-        statsWindow.entries.some(
-          (candidate) =>
-            candidate !== entry &&
-            candidate.statId === entry.statId &&
-            candidate.targetId === nextTargetId,
-        )
-      ) {
-        renderStatsWindow(statsWindow);
-        return;
-      }
-      const index = statsWindow.entries.findIndex((candidate) => candidate.key === entry.key);
-      if (index >= 0) {
-        statsWindow.entries[index] = {
-          key: `${statsWindow.id}:${entry.statId}:${nextTargetId}`,
-          statId: entry.statId,
-          targetId: nextTargetId,
-        };
-      }
-      renderStatsWindow(statsWindow);
-      scheduleConfigUrlUpdate();
-    });
-    name.append(" ", targetSelect);
-  }
-  const valueEl = document.createElement("span");
-  valueEl.className = "stats-window-stat-value";
-  valueEl.textContent = value;
-  const remove = document.createElement("button");
-  remove.type = "button";
-  remove.className = "stats-window-stat-remove";
-  remove.textContent = "x";
-  remove.addEventListener("click", () => {
-    removeStatFromWindow(statsWindow, entry.key);
-    renderStatsWindow(statsWindow);
-    scheduleConfigUrlUpdate();
-  });
-  row.append(name, valueEl, remove);
-  return row;
-}
-
-function formatSetting(value: number | undefined, suffix = "", digits = 0): string {
-  if (value === undefined || Number.isNaN(value)) {
-    return "--";
-  }
-
-  return `${value.toFixed(digits)}${suffix}`;
-}
-
-function getFallbackCameraSettings(): Required<CameraSettings> {
-  return {
-    fov: 110,
-    height: 100,
-    pitch: -4,
-    distance: 270,
-    stiffness: 0,
-    swivelSpeed: 1,
-    transitionSpeed: 1,
-  };
-}
-
-function getAttachedPlayerCameraSettings(attachedPlayerId: string | null): CameraSettings | null {
-  if (!replayPlayer || attachedPlayerId === null) {
-    return null;
-  }
-
-  return (
-    replayPlayer.replay.players.find((candidate) => candidate.id === attachedPlayerId)
-      ?.cameraSettings ?? null
-  );
-}
-
-function getEffectiveCameraSettings(state: ReplayPlayerState): CameraSettings {
-  return {
-    ...getFallbackCameraSettings(),
-    ...(getAttachedPlayerCameraSettings(state.attachedPlayerId) ?? {}),
-    ...(state.customCameraSettings ?? {}),
-  };
-}
-
-function readCustomCameraSettings(): CameraSettings {
-  return {
-    fov: Number(customCameraFov.value),
-    height: Number(customCameraHeight.value),
-    pitch: Number(customCameraPitch.value),
-    distance: Number(customCameraDistance.value),
-    stiffness: Number(customCameraStiffness.value),
-    swivelSpeed: Number(customCameraSwivelSpeed.value),
-    transitionSpeed: Number(customCameraTransitionSpeed.value),
-  };
-}
-
-function setCameraSettingControlsEnabled(enabled: boolean): void {
-  cameraSettingsControls.hidden = !customCameraSettings.checked;
-  customCameraFov.disabled = !enabled;
-  customCameraHeight.disabled = !enabled;
-  customCameraPitch.disabled = !enabled;
-  customCameraDistance.disabled = !enabled;
-  customCameraStiffness.disabled = !enabled;
-  customCameraSwivelSpeed.disabled = !enabled;
-  customCameraTransitionSpeed.disabled = !enabled;
-}
-
-function syncCustomCameraSettingControls(settings: CameraSettings): void {
-  const fallback = getFallbackCameraSettings();
-  const fov = settings.fov ?? fallback.fov;
-  const height = settings.height ?? fallback.height;
-  const pitch = settings.pitch ?? fallback.pitch;
-  const distance = settings.distance ?? fallback.distance;
-  const stiffness = settings.stiffness ?? fallback.stiffness;
-  const swivelSpeed = settings.swivelSpeed ?? fallback.swivelSpeed;
-  const transitionSpeed = settings.transitionSpeed ?? fallback.transitionSpeed;
-
-  customCameraFov.value = `${fov}`;
-  customCameraHeight.value = `${height}`;
-  customCameraPitch.value = `${pitch}`;
-  customCameraDistance.value = `${distance}`;
-  customCameraStiffness.value = `${stiffness}`;
-  customCameraSwivelSpeed.value = `${swivelSpeed}`;
-  customCameraTransitionSpeed.value = `${transitionSpeed}`;
-
-  customCameraFovReadout.textContent = formatSetting(fov, "", 0);
-  customCameraHeightReadout.textContent = formatSetting(height, "", 0);
-  customCameraPitchReadout.textContent = formatSetting(pitch, "", 0);
-  customCameraDistanceReadout.textContent = formatSetting(distance, "", 0);
-  customCameraStiffnessReadout.textContent = formatSetting(stiffness, "", 2);
-  customCameraSwivelSpeedReadout.textContent = formatSetting(swivelSpeed, "", 1);
-  customCameraTransitionSpeedReadout.textContent = formatSetting(transitionSpeed, "", 2);
-}
-
 function setTransportEnabled(enabled: boolean): void {
   togglePlayback.disabled = !enabled;
   playbackRate.disabled = !enabled;
-  attachedPlayer.disabled = !enabled;
   skipPostGoalTransitions.disabled = !enabled;
   skipKickoffs.disabled = !enabled;
   hitboxWireframes.disabled = !enabled;
-  syncCameraModeButtons(enabled ? replayPlayer?.getState() : undefined);
-}
-
-function getCameraViewButton(mode: ReplayCameraViewMode): HTMLButtonElement {
-  switch (mode) {
-    case "free":
-      return cameraViewFreeButton;
-    case "follow":
-      return cameraViewFollowButton;
-  }
-}
-
-function syncCameraModeButtons(state?: ReplayPlayerState): void {
-  const activeMode = state?.cameraViewMode ?? "free";
-  const hasReplay = replayPlayer !== null && state !== undefined;
-  const canFollow = (state?.attachedPlayerId ?? null) !== null;
-
-  for (const mode of CAMERA_VIEW_MODES) {
-    const button = getCameraViewButton(mode);
-    button.disabled = !hasReplay || (mode === "follow" && !canFollow);
-    const isActive = mode === activeMode;
-    button.dataset.active = isActive ? "true" : "false";
-    button.setAttribute("aria-pressed", isActive ? "true" : "false");
-  }
-
-  cameraViewOverheadButton.disabled = !hasReplay;
-  cameraViewSideButton.disabled = !hasReplay;
-  cameraViewOverheadButton.dataset.active = "false";
-  cameraViewSideButton.dataset.active = "false";
-  cameraViewOverheadButton.setAttribute("aria-pressed", "false");
-  cameraViewSideButton.setAttribute("aria-pressed", "false");
-}
-
-function syncCameraControlAvailability(state?: ReplayPlayerState): void {
-  syncCameraModeButtons(state);
-  const hasAttachedCamera =
-    replayPlayer !== null &&
-    state?.cameraViewMode === "follow" &&
-    (state.attachedPlayerId ?? null) !== null;
-  cameraDistance.disabled = !hasAttachedCamera;
-  customCameraSettings.disabled = !hasAttachedCamera;
-  setCameraSettingControlsEnabled(hasAttachedCamera && state?.customCameraSettings !== null);
-  ballCam.disabled = !hasAttachedCamera;
-}
-
-function populateAttachedPlayerOptions(players: ReplayPlayerTrack[]): void {
-  attachedPlayer.replaceChildren();
-  attachedPlayer.append(new Option("Free camera", ""));
-
-  for (const player of players) {
-    attachedPlayer.append(
-      new Option(`${player.name} (${player.isTeamZero ? "Blue" : "Orange"})`, player.id),
-    );
-  }
-}
-
-function formatBytes(bytes: number): string {
-  if (bytes <= 0) {
-    return "--";
-  }
-  const units = ["B", "KB", "MB", "GB"];
-  let value = bytes;
-  let unitIndex = 0;
-  while (value >= 1024 && unitIndex < units.length - 1) {
-    value /= 1024;
-    unitIndex += 1;
-  }
-  const precision = unitIndex === 0 ? 0 : value >= 10 ? 1 : 2;
-  return `${value.toFixed(precision)} ${units[unitIndex]}`;
-}
-
-function recordingLabel(status: CanvasRecorderStatus | null): string {
-  if (!status) {
-    return "No replay";
-  }
-  if (status.error) {
-    return status.error;
-  }
-  switch (status.state) {
-    case "idle":
-      return "Idle";
-    case "recording":
-      return "Recording";
-    case "stopping":
-      return "Stopping";
-    case "ready":
-      return "Ready";
-    case "error":
-      return "Error";
-  }
-}
-
-function getRecordingOptions(): { fps: number; playbackRate: number } {
-  const fps = Number(recordingFps.value);
-  const playbackRate = Number(recordingPlaybackRate.value);
-  return {
-    fps: Number.isFinite(fps) ? Math.max(1, Math.min(120, Math.trunc(fps))) : 60,
-    playbackRate: Number.isFinite(playbackRate) ? Math.max(0.1, playbackRate) : 1,
-  };
+  cameraControlsController?.setTransportEnabled(enabled, replayPlayer?.getState());
 }
 
 function syncRecordingWindow(status = canvasRecorder?.getStatus() ?? null): void {
-  const hasRecorder = canvasRecorder !== null && replayPlayer !== null;
-  const state = status?.state ?? "idle";
-  const isRecording = state === "recording" || state === "stopping";
-  const hasRecording = (canvasRecorder?.getRecording() ?? null) !== null;
-
-  recordingStatus.textContent = recordingLabel(status);
-  recordingElapsed.textContent = `${(status?.elapsedSeconds ?? 0).toFixed(1)}s`;
-  recordingSize.textContent = formatBytes(status?.sizeBytes ?? 0);
-  recordingType.textContent = status?.mimeType || "WebM";
-  recordingStart.disabled = !hasRecorder || isRecording;
-  recordingFullReplay.disabled = !hasRecorder || isRecording;
-  recordingStop.disabled = !hasRecorder || !isRecording;
-  recordingDownload.disabled = !hasRecording || isRecording;
-  recordingClear.disabled = !hasRecording || isRecording;
-  recordingFps.disabled = isRecording;
-  recordingPlaybackRate.disabled = isRecording;
-}
-
-function recordingFileName(): string {
-  const source = loadedReplayName?.replace(/\.replay$/i, "") || "replay";
-  const safeSource = source.replace(/[^a-zA-Z0-9._-]+/g, "-").replace(/^-+|-+$/g, "");
-  const timestamp = new Date().toISOString().replace(/[:.]/g, "-");
-  return `${safeSource || "replay"}-${timestamp}.webm`;
-}
-
-function downloadRecording(blob: Blob): void {
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = recordingFileName();
-  document.body.append(link);
-  link.click();
-  link.remove();
-  window.setTimeout(() => URL.revokeObjectURL(url), 0);
-}
-
-function renderCameraProfile(state?: ReplayPlayerState): void {
-  const attachedPlayerId = state?.attachedPlayerId ?? null;
-  if (!replayPlayer || state?.cameraViewMode !== "follow" || attachedPlayerId === null) {
-    cameraProfileReadout.textContent = "Free camera";
-    cameraFovReadout.textContent = "--";
-    cameraHeightReadout.textContent = "--";
-    cameraPitchReadout.textContent = "--";
-    cameraBaseDistanceReadout.textContent = "--";
-    cameraStiffnessReadout.textContent = "--";
-    return;
-  }
-
-  const player = replayPlayer.replay.players.find((candidate) => candidate.id === attachedPlayerId);
-  if (!player) {
-    cameraProfileReadout.textContent = "Unknown";
-    cameraFovReadout.textContent = "--";
-    cameraHeightReadout.textContent = "--";
-    cameraPitchReadout.textContent = "--";
-    cameraBaseDistanceReadout.textContent = "--";
-    cameraStiffnessReadout.textContent = "--";
-    return;
-  }
-
-  const cameraSettings = getEffectiveCameraSettings(state);
-  cameraProfileReadout.textContent =
-    state.customCameraSettings === null ? player.name : `${player.name} custom`;
-  cameraFovReadout.textContent = formatSetting(cameraSettings.fov, "", 0);
-  cameraHeightReadout.textContent = formatSetting(cameraSettings.height, "", 0);
-  cameraPitchReadout.textContent = formatSetting(cameraSettings.pitch, "", 0);
-  cameraBaseDistanceReadout.textContent = formatSetting(cameraSettings.distance, "", 0);
-  cameraStiffnessReadout.textContent = formatSetting(cameraSettings.stiffness, "", 2);
+  recordingWindowController?.sync(status);
 }
 
 function renderSnapshot(state: ReplayPlayerState): void {
@@ -3397,20 +2045,12 @@ function renderSnapshot(state: ReplayPlayerState): void {
   playbackStatusReadout.textContent = state.playing ? "Playing" : "Paused";
   togglePlayback.textContent = state.playing ? "Pause" : "Play";
   playbackRate.value = `${state.speed}`;
-  cameraDistance.value = `${state.cameraDistanceScale}`;
-  cameraDistanceReadout.textContent = `${state.cameraDistanceScale.toFixed(2)}x`;
-  customCameraSettings.checked = state.customCameraSettings !== null;
-  cameraSettingsControls.hidden = !customCameraSettings.checked;
-  syncCustomCameraSettingControls(getEffectiveCameraSettings(state));
-  ballCam.checked = state.ballCamEnabled;
-  attachedPlayer.value = state.attachedPlayerId ?? "";
+  cameraControlsController?.syncState(state);
   skipPostGoalTransitions.checked = state.skipPostGoalTransitionsEnabled;
   skipKickoffs.checked = state.skipKickoffsEnabled;
   hitboxWireframes.checked = state.hitboxWireframesEnabled;
   emptyState.hidden = true;
 
-  syncCameraControlAvailability(state);
-  renderCameraProfile(state);
   renderStatsWindows(state.frameIndex, { preserveOpenPickers: true });
   renderScoreboard(state.frameIndex);
   syncEventPlaylistTimeline(state);
@@ -3488,7 +2128,7 @@ async function loadReplayBundleForDisplay(
   fileInput.disabled = true;
   replayLoadModal?.show(source.name, source.preparingStatus);
   setTransportEnabled(false);
-  syncCameraControlAvailability();
+  cameraControlsController?.syncAvailability();
   emptyState.hidden = false;
 
   if (unsubscribe) {
@@ -3509,8 +2149,7 @@ async function loadReplayBundleForDisplay(
   clearTimelineRangeSources();
   clearStandalonePlugins();
   clearRenderCaches();
-  eventPlaylistActiveSourceIds = null;
-  eventPlaylistLastActiveKey = null;
+  resetEventPlaylistWindow();
   renderScoreboard();
   renderTimelineEventCount();
   renderMechanicsTimelineControls();
@@ -3574,7 +2213,7 @@ async function loadReplayBundleForDisplay(
       }
     }
 
-    populateAttachedPlayerOptions(replay.players);
+    cameraControlsController?.populateAttachedPlayerOptions(replay.players);
     emptyState.hidden = true;
     statusReadout.textContent = `Loaded ${source.name}`;
     loadedReplayName = source.name;
@@ -3582,11 +2221,10 @@ async function loadReplayBundleForDisplay(
     framesReadout.textContent = `${replay.frameCount}`;
     renderTimelineEventCount();
     renderMechanicsTimelineControls();
-    eventPlaylistActiveSourceIds = null;
-    eventPlaylistLastActiveKey = null;
+    resetEventPlaylistWindow();
     renderEventPlaylistWindow();
     setTransportEnabled(true);
-    syncCameraControlAvailability(replayPlayer.getState());
+    cameraControlsController?.syncAvailability(replayPlayer.getState());
     renderSnapshot(replayPlayer.getState());
     renderStatsWindows(replayPlayer.getState().frameIndex);
     renderScoreboard(replayPlayer.getState().frameIndex);
@@ -3654,6 +2292,13 @@ export function mountStatEvaluationPlayer(
     "#mechanics-timeline-window-body",
   );
   eventPlaylistWindowBody = mustElement<HTMLDivElement>(root, "#event-playlist-window-body");
+  eventPlaylistController = createEventPlaylistWindowController({
+    body: eventPlaylistWindowBody,
+    getReplayPlayer: () => replayPlayer,
+    getSources: getEventPlaylistSourcesForWindow,
+    cueTimelineEvent,
+    formatTime,
+  });
   mechanicsReviewFile = mustElement<HTMLInputElement>(root, "#mechanics-review-file");
   mechanicsReviewUrl = mustElement<HTMLInputElement>(root, "#mechanics-review-url");
   mechanicsReviewLoadUrl = mustElement<HTMLButtonElement>(root, "#mechanics-review-load-url");
@@ -3686,41 +2331,73 @@ export function mountStatEvaluationPlayer(
   );
   touchControlsWindowBody = mustElement<HTMLDivElement>(root, "#touch-controls-window-body");
   statsWindowLayer = mustElement<HTMLDivElement>(root, "#stats-window-layer");
+  statsWindowsController = createStatsWindowsController({
+    layer: statsWindowLayer,
+    getReplayPlayer: () => replayPlayer,
+    getStatsTimeline: () => statsTimeline,
+    getStatsFrameLookup: () => statsFrameLookup,
+    getStatRegistry: () => statRegistry,
+    readWindowPlacement,
+    applyWindowPlacement,
+    bringWindowToFront,
+    setLauncherOpen,
+    requestConfigSync: scheduleConfigUrlUpdate,
+    watchGoalReplay,
+    cueGoalReplay,
+  });
   togglePlayback = mustElement<HTMLButtonElement>(root, "#toggle-playback");
   playbackRate = mustElement<HTMLSelectElement>(root, "#playback-rate");
-  attachedPlayer = mustElement<HTMLSelectElement>(root, "#attached-player");
-  cameraViewFreeButton = mustElement<HTMLButtonElement>(root, "#camera-view-free");
-  cameraViewFollowButton = mustElement<HTMLButtonElement>(root, "#camera-view-follow");
-  cameraViewOverheadButton = mustElement<HTMLButtonElement>(root, "#camera-view-overhead");
-  cameraViewSideButton = mustElement<HTMLButtonElement>(root, "#camera-view-side");
-  cameraDistance = mustElement<HTMLInputElement>(root, "#camera-distance");
-  cameraDistanceReadout = mustElement<HTMLElement>(root, "#camera-distance-readout");
-  customCameraSettings = mustElement<HTMLInputElement>(root, "#custom-camera-settings");
-  cameraSettingsControls = mustElement<HTMLDivElement>(root, "#camera-settings-controls");
-  customCameraFov = mustElement<HTMLInputElement>(root, "#custom-camera-fov");
-  customCameraHeight = mustElement<HTMLInputElement>(root, "#custom-camera-height");
-  customCameraPitch = mustElement<HTMLInputElement>(root, "#custom-camera-pitch");
-  customCameraDistance = mustElement<HTMLInputElement>(root, "#custom-camera-distance");
-  customCameraStiffness = mustElement<HTMLInputElement>(root, "#custom-camera-stiffness");
-  customCameraSwivelSpeed = mustElement<HTMLInputElement>(root, "#custom-camera-swivel-speed");
-  customCameraTransitionSpeed = mustElement<HTMLInputElement>(
-    root,
-    "#custom-camera-transition-speed",
-  );
-  customCameraFovReadout = mustElement<HTMLElement>(root, "#custom-camera-fov-readout");
-  customCameraHeightReadout = mustElement<HTMLElement>(root, "#custom-camera-height-readout");
-  customCameraPitchReadout = mustElement<HTMLElement>(root, "#custom-camera-pitch-readout");
-  customCameraDistanceReadout = mustElement<HTMLElement>(root, "#custom-camera-distance-readout");
-  customCameraStiffnessReadout = mustElement<HTMLElement>(root, "#custom-camera-stiffness-readout");
-  customCameraSwivelSpeedReadout = mustElement<HTMLElement>(
-    root,
-    "#custom-camera-swivel-speed-readout",
-  );
-  customCameraTransitionSpeedReadout = mustElement<HTMLElement>(
-    root,
-    "#custom-camera-transition-speed-readout",
-  );
-  ballCam = mustElement<HTMLInputElement>(root, "#ball-cam");
+  cameraControlsController = createCameraControlsController({
+    elements: {
+      attachedPlayer: mustElement<HTMLSelectElement>(root, "#attached-player"),
+      cameraViewFreeButton: mustElement<HTMLButtonElement>(root, "#camera-view-free"),
+      cameraViewFollowButton: mustElement<HTMLButtonElement>(root, "#camera-view-follow"),
+      cameraViewOverheadButton: mustElement<HTMLButtonElement>(root, "#camera-view-overhead"),
+      cameraViewSideButton: mustElement<HTMLButtonElement>(root, "#camera-view-side"),
+      cameraDistance: mustElement<HTMLInputElement>(root, "#camera-distance"),
+      cameraDistanceReadout: mustElement<HTMLElement>(root, "#camera-distance-readout"),
+      customCameraSettings: mustElement<HTMLInputElement>(root, "#custom-camera-settings"),
+      cameraSettingsControls: mustElement<HTMLDivElement>(root, "#camera-settings-controls"),
+      customCameraFov: mustElement<HTMLInputElement>(root, "#custom-camera-fov"),
+      customCameraHeight: mustElement<HTMLInputElement>(root, "#custom-camera-height"),
+      customCameraPitch: mustElement<HTMLInputElement>(root, "#custom-camera-pitch"),
+      customCameraDistance: mustElement<HTMLInputElement>(root, "#custom-camera-distance"),
+      customCameraStiffness: mustElement<HTMLInputElement>(root, "#custom-camera-stiffness"),
+      customCameraSwivelSpeed: mustElement<HTMLInputElement>(root, "#custom-camera-swivel-speed"),
+      customCameraTransitionSpeed: mustElement<HTMLInputElement>(
+        root,
+        "#custom-camera-transition-speed",
+      ),
+      customCameraFovReadout: mustElement<HTMLElement>(root, "#custom-camera-fov-readout"),
+      customCameraHeightReadout: mustElement<HTMLElement>(root, "#custom-camera-height-readout"),
+      customCameraPitchReadout: mustElement<HTMLElement>(root, "#custom-camera-pitch-readout"),
+      customCameraDistanceReadout: mustElement<HTMLElement>(
+        root,
+        "#custom-camera-distance-readout",
+      ),
+      customCameraStiffnessReadout: mustElement<HTMLElement>(
+        root,
+        "#custom-camera-stiffness-readout",
+      ),
+      customCameraSwivelSpeedReadout: mustElement<HTMLElement>(
+        root,
+        "#custom-camera-swivel-speed-readout",
+      ),
+      customCameraTransitionSpeedReadout: mustElement<HTMLElement>(
+        root,
+        "#custom-camera-transition-speed-readout",
+      ),
+      ballCam: mustElement<HTMLInputElement>(root, "#ball-cam"),
+      cameraProfileReadout: mustElement<HTMLElement>(root, "#camera-profile-readout"),
+      cameraFovReadout: mustElement<HTMLElement>(root, "#camera-fov-readout"),
+      cameraHeightReadout: mustElement<HTMLElement>(root, "#camera-height-readout"),
+      cameraPitchReadout: mustElement<HTMLElement>(root, "#camera-pitch-readout"),
+      cameraBaseDistanceReadout: mustElement<HTMLElement>(root, "#camera-base-distance-readout"),
+      cameraStiffnessReadout: mustElement<HTMLElement>(root, "#camera-stiffness-readout"),
+    },
+    getReplayPlayer: () => replayPlayer,
+    requestConfigSync: scheduleConfigUrlUpdate,
+  });
   moduleSummaryEl = mustElement<HTMLDivElement>(root, "#module-summary");
   moduleSettingsEl = mustElement<HTMLDivElement>(root, "#module-settings");
   timeReadout = mustElement<HTMLElement>(root, "#time-readout");
@@ -3731,26 +2408,31 @@ export function mountStatEvaluationPlayer(
   playersReadout = mustElement<HTMLElement>(root, "#players-readout");
   framesReadout = mustElement<HTMLElement>(root, "#frames-readout");
   eventsReadout = mustElement<HTMLElement>(root, "#events-readout");
-  cameraProfileReadout = mustElement<HTMLElement>(root, "#camera-profile-readout");
-  cameraFovReadout = mustElement<HTMLElement>(root, "#camera-fov-readout");
-  cameraHeightReadout = mustElement<HTMLElement>(root, "#camera-height-readout");
-  cameraPitchReadout = mustElement<HTMLElement>(root, "#camera-pitch-readout");
-  cameraBaseDistanceReadout = mustElement<HTMLElement>(root, "#camera-base-distance-readout");
-  cameraStiffnessReadout = mustElement<HTMLElement>(root, "#camera-stiffness-readout");
   skipPostGoalTransitions = mustElement<HTMLInputElement>(root, "#skip-post-goal-transitions");
   skipKickoffs = mustElement<HTMLInputElement>(root, "#skip-kickoffs");
   hitboxWireframes = mustElement<HTMLInputElement>(root, "#hitbox-wireframes");
-  recordingFps = mustElement<HTMLInputElement>(root, "#recording-fps");
-  recordingPlaybackRate = mustElement<HTMLSelectElement>(root, "#recording-playback-rate");
-  recordingStart = mustElement<HTMLButtonElement>(root, "#recording-start");
-  recordingFullReplay = mustElement<HTMLButtonElement>(root, "#recording-full-replay");
-  recordingStop = mustElement<HTMLButtonElement>(root, "#recording-stop");
-  recordingDownload = mustElement<HTMLButtonElement>(root, "#recording-download");
-  recordingClear = mustElement<HTMLButtonElement>(root, "#recording-clear");
-  recordingStatus = mustElement<HTMLElement>(root, "#recording-status");
-  recordingElapsed = mustElement<HTMLElement>(root, "#recording-elapsed");
-  recordingSize = mustElement<HTMLElement>(root, "#recording-size");
-  recordingType = mustElement<HTMLElement>(root, "#recording-type");
+  recordingWindowController = createRecordingWindowController({
+    elements: {
+      fps: mustElement<HTMLInputElement>(root, "#recording-fps"),
+      playbackRate: mustElement<HTMLSelectElement>(root, "#recording-playback-rate"),
+      start: mustElement<HTMLButtonElement>(root, "#recording-start"),
+      fullReplay: mustElement<HTMLButtonElement>(root, "#recording-full-replay"),
+      stop: mustElement<HTMLButtonElement>(root, "#recording-stop"),
+      download: mustElement<HTMLButtonElement>(root, "#recording-download"),
+      clear: mustElement<HTMLButtonElement>(root, "#recording-clear"),
+      status: mustElement<HTMLElement>(root, "#recording-status"),
+      elapsed: mustElement<HTMLElement>(root, "#recording-elapsed"),
+      size: mustElement<HTMLElement>(root, "#recording-size"),
+      type: mustElement<HTMLElement>(root, "#recording-type"),
+    },
+    getCanvasRecorder: () => canvasRecorder,
+    getReplayPlayer: () => replayPlayer,
+    getLoadedReplayName: () => loadedReplayName,
+    setStatus(message) {
+      statusReadout.textContent = message;
+    },
+    requestConfigSync: scheduleConfigUrlUpdate,
+  });
 
   const configParamSnapshot = getStatsPlayerConfigParamSnapshot(window.location);
   const configDebugEnabled = isStatsPlayerConfigDebugEnabled(window.location);
@@ -3787,7 +2469,8 @@ export function mountStatEvaluationPlayer(
     statsTimeline = null;
     statsFrameLookup = null;
     statRegistry = createStatRegistry(null);
-    statsWindows.clear();
+    clearStatsWindows();
+    statsWindowsController = null;
     clearTimelineEventSources();
     clearTimelineRangeSources();
     clearStandalonePlugins();
@@ -3799,21 +2482,20 @@ export function mountStatEvaluationPlayer(
     activeTimelineRangeModuleIds = new Set<string>();
     activeMechanicTimelineKinds = new Set<string>();
     activeRenderEffectModuleIds = new Set<string>();
-    eventPlaylistActiveSourceIds = null;
-    eventPlaylistAutoFollow = true;
-    eventPlaylistLastActiveKey = null;
+    resetEventPlaylistWindow();
+    eventPlaylistController = null;
     activeMechanicsReview = null;
     mechanicsReviewBoundaryGuard = false;
     boostPadOverlayEnabled = true;
     loadedReplayName = null;
-    lastFreeCameraPreset = null;
+    cameraControlsController = null;
+    recordingWindowController = null;
     initialUrlConfig = null;
     if (configUrlUpdateTimer !== null) {
       window.clearTimeout(configUrlUpdateTimer);
       configUrlUpdateTimer = null;
     }
     isApplyingConfig = false;
-    nextStatsWindowId = 1;
     nextWindowZIndex = 30;
     removeRenderHook = null;
     if (appRoot === root) {
@@ -4026,194 +2708,8 @@ export function mountStatEvaluationPlayer(
     { signal: listeners.signal },
   );
 
-  recordingStart.addEventListener(
-    "click",
-    () => {
-      if (!canvasRecorder) {
-        return;
-      }
-      try {
-        const { fps } = getRecordingOptions();
-        canvasRecorder.start({ fps });
-        syncRecordingWindow();
-      } catch (error) {
-        console.error("Failed to start recording:", error);
-        statusReadout.textContent =
-          error instanceof Error ? error.message : "Failed to start recording";
-        syncRecordingWindow(canvasRecorder.getStatus());
-      }
-    },
-    { signal: listeners.signal },
-  );
-
-  recordingFullReplay.addEventListener(
-    "click",
-    () => {
-      if (!canvasRecorder) {
-        return;
-      }
-      const { fps, playbackRate } = getRecordingOptions();
-      void canvasRecorder
-        .recordFullReplay({
-          fps,
-          playbackRate,
-          restorePlaybackState: true,
-        })
-        .catch((error) => {
-          console.error("Failed to record replay:", error);
-          statusReadout.textContent =
-            error instanceof Error ? error.message : "Failed to record replay";
-          syncRecordingWindow(canvasRecorder?.getStatus() ?? null);
-        });
-      syncRecordingWindow();
-    },
-    { signal: listeners.signal },
-  );
-
-  recordingStop.addEventListener(
-    "click",
-    () => {
-      void canvasRecorder?.stop().catch((error) => {
-        console.error("Failed to stop recording:", error);
-        statusReadout.textContent =
-          error instanceof Error ? error.message : "Failed to stop recording";
-      });
-      syncRecordingWindow();
-    },
-    { signal: listeners.signal },
-  );
-
-  recordingDownload.addEventListener(
-    "click",
-    () => {
-      const blob = canvasRecorder?.getRecording();
-      if (blob) {
-        downloadRecording(blob);
-      }
-    },
-    { signal: listeners.signal },
-  );
-
-  recordingClear.addEventListener(
-    "click",
-    () => {
-      try {
-        canvasRecorder?.clear();
-        syncRecordingWindow();
-      } catch (error) {
-        console.error("Failed to clear recording:", error);
-      }
-    },
-    { signal: listeners.signal },
-  );
-
-  recordingFps.addEventListener("change", scheduleConfigUrlUpdate, {
-    signal: listeners.signal,
-  });
-  recordingPlaybackRate.addEventListener("change", scheduleConfigUrlUpdate, {
-    signal: listeners.signal,
-  });
-
-  cameraDistance.addEventListener(
-    "input",
-    () => {
-      replayPlayer?.setCameraDistanceScale(Number(cameraDistance.value));
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  customCameraSettings.addEventListener(
-    "change",
-    () => {
-      cameraSettingsControls.hidden = !customCameraSettings.checked;
-      replayPlayer?.setCustomCameraSettings(
-        customCameraSettings.checked ? readCustomCameraSettings() : null,
-      );
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  for (const input of [
-    customCameraFov,
-    customCameraHeight,
-    customCameraPitch,
-    customCameraDistance,
-    customCameraStiffness,
-    customCameraSwivelSpeed,
-    customCameraTransitionSpeed,
-  ]) {
-    input.addEventListener(
-      "input",
-      () => {
-        const settings = readCustomCameraSettings();
-        syncCustomCameraSettingControls(settings);
-        replayPlayer?.setCustomCameraSettings(settings);
-        scheduleConfigUrlUpdate();
-      },
-      { signal: listeners.signal },
-    );
-  }
-
-  attachedPlayer.addEventListener(
-    "change",
-    () => {
-      replayPlayer?.setAttachedPlayer(attachedPlayer.value || null);
-      lastFreeCameraPreset = null;
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  cameraViewFreeButton.addEventListener(
-    "click",
-    () => {
-      replayPlayer?.setCameraViewMode("free");
-      lastFreeCameraPreset = null;
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  cameraViewFollowButton.addEventListener(
-    "click",
-    () => {
-      replayPlayer?.setCameraViewMode("follow");
-      lastFreeCameraPreset = null;
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  cameraViewOverheadButton.addEventListener(
-    "click",
-    () => {
-      replayPlayer?.setFreeCameraPreset("overhead");
-      lastFreeCameraPreset = "overhead";
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  cameraViewSideButton.addEventListener(
-    "click",
-    () => {
-      replayPlayer?.setFreeCameraPreset("side");
-      lastFreeCameraPreset = "side";
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
-
-  ballCam.addEventListener(
-    "change",
-    () => {
-      replayPlayer?.setBallCamEnabled(ballCam.checked);
-      scheduleConfigUrlUpdate();
-    },
-    { signal: listeners.signal },
-  );
+  recordingWindowController?.installEventListeners(listeners.signal);
+  cameraControlsController?.installEventListeners(listeners.signal);
 
   skipPostGoalTransitions.addEventListener(
     "change",
@@ -4245,8 +2741,8 @@ export function mountStatEvaluationPlayer(
   renderModuleSummary();
   renderModuleSettings();
   renderScoreboard();
-  renderCameraProfile();
-  syncCameraModeButtons();
+  cameraControlsController?.renderProfile();
+  cameraControlsController?.syncModeButtons();
   syncRecordingWindow();
   renderTimelineEventCount();
   renderMechanicsReviewWindow();
